requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.0
Flask==3.0.0
flask-cors-4.0.0
fonttools==4.43.0
frozendict==2.3.8
html5lib==1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.0
multitasking==0.0.11
numpy==1.26.0
packaging==23.2
pandas==2.1.1
patsy==0.5.3
peewee==3.16.3
Pillow==10.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scipy==1.11.3
six==1.16.0
soupsieve==2.5
statsmodels==0.14.0
tzdata==2023.3
urllib3==2.0.6
webencodings==0.5.1
Werkzeug==3.0.0
yfinance==0.2.30
